Initial modifications for spectrum flipping Make frequency domain bins always monotonically increase by flipping every other sample of complex A/D inputs and of real inputs from inverted spectra (tuners with high side LO injection)

+++ rx888.c
@@ -285,7 +285,7 @@ int rx888_setup(struct frontend * const frontend,dictionary const * const dictio
     reference = Default_reference;
   }
   double calibrate = config_getdouble(dictionary,section,"calibrate",0);
-  if(fabsl(calibrate) >= Max_calibrate){
+  if(fabs(calibrate) >= Max_calibrate){
     fprintf(stderr,"Unreasonable frequency calibration %.3g, setting to 0\n",calibrate);
     calibrate = 0;
   }
